Casts for result of strlen to int to fix MSVC builds (size_t is bigger than int)

+++ main.c
@@ -157,7 +157,7 @@ int main(int argc, char *argv[]) {
     char title[] = "WLA " ARCH_STR " Macro Assembler v10.1a";
     int length, left, right;
 
-    length = strlen(title);
+    length = (int)strlen(title);
     left = (70 - 3 - 3 - length) / 2;
     right = 70 - 3 - 3 - left - length;
 
@@ -173,7 +173,7 @@ int main(int argc, char *argv[]) {
     printf("                Programmed by Ville Helin in 1998-2008\n");
     printf("        In GitHub since 2014: https://github.com/vhelin/wla-dx\n");
 
-    length = strlen(g_version_string);
+    length = (int)strlen(g_version_string);
     left = (70 - length) / 2;
 
     for (q = 0; q < left; q++)
